Add cast to hdr->l_text in dump_hdr()
Avoids warning when the htslib header API gets merged.

+++ test/test.c
@@ -67,7 +67,7 @@ void dump_hdr(const bam_hdr_t* hdr)
 {
     printf("n_targets: %d\n", hdr->n_targets);
     printf("ignore_sam_err: %d\n", hdr->ignore_sam_err);
-    printf("l_text: %u\n", hdr->l_text);
+    printf("l_text: %zu\n", (size_t) hdr->l_text);
     printf("idx\ttarget_len\ttarget_name:\n");
     int32_t target;
     for (target = 0; target < hdr->n_targets; ++target) {
